[feat] 운동 데이터 입력

+++ src/main/java/com/feedback/hafit/domain/exercise/dto/request/ExerciseRequestDTO.java
@@ -14,7 +14,7 @@ public class ExerciseRequestDTO {
 
     private String exercise_name; // 이름
 
-    private Long exercise_calorie; // 칼로리
+    private int exercise_calorie; // 칼로리
 
     private String exercise_description; // 설명
 
@@ -28,7 +28,7 @@ public class ExerciseRequestDTO {
         this.exercise_name = exercise_name;
     }
 
-    public void setExercise_calorie(Long exercise_calorie) {
+    public void setExercise_calorie(int exercise_calorie) {
         this.exercise_calorie = exercise_calorie;
     }
